chore: move "base_model_factory" folder to "domain"
