+++ lib/domain/audio/factories/audio_info_set_factory.dart
@@ -2,7 +2,7 @@ import 'package:dune/domain/audio/base_models/audio_info_set.dart';
 import 'package:dune/domain/audio/base_models/track_audio_info.dart';
 import 'package:dune/support/enums/music_source.dart';
 
-import 'base_model_factory.dart';
+import '../../base_model_factory.dart';
 import 'track_audio_info_factory.dart';
 
 final class AudioInfoSetFactory extends BaseModelFactory<AudioInfoSet> {
