+++ lib/domain/audio/factories/date_listening_history_factory.dart
@@ -2,7 +2,7 @@ import 'package:dune/domain/audio/base_models/base_playlists_listening_history.d
 import 'package:dune/domain/audio/base_models/base_track.dart';
 import 'package:dune/domain/audio/base_models/base_track_listening_history.dart';
 import 'package:dune/domain/audio/base_models/listening_history_collection.dart';
-import 'package:dune/domain/audio/factories/base_model_factory.dart';
+import 'package:dune/domain/base_model_factory.dart';
 import 'package:dune/support/extensions/extensions.dart';
 import 'package:faker/faker.dart';
