+++ lib/domain/audio/factories/explore_music_collection_factory.dart
@@ -1,6 +1,6 @@
 import 'package:dune/domain/audio/base_models/base_explore_music_collection.dart';
 import 'package:dune/domain/audio/base_models/base_explore_music_item.dart';
-import 'package:dune/domain/audio/factories/base_model_factory.dart';
+import 'package:dune/domain/base_model_factory.dart';
 import 'package:dune/support/enums/music_source.dart';
 import 'package:faker/faker.dart';
